publish data and wb in private
publish data and wb in private
